Lesson 7: Product & Rating components, compute stars based on rating value.
We extract the product and rating code from App.js into separate react
components Product.js and Rating.js, and use the <Product> component in App.js
to show the product card, and within the product component, we use the <Rating>
component to display the rating stars. Now we change the css classes of the
stars so that they display a number of solid stars based on the product rating
value.