Lesson 8: Build Home (product list) & Product screens.
We used npm to install react-router-dom for routing in the frontend react app.
We created two new React components, HomeScreen and ProductScreen. We are
commiting the package-lock.json file to the git repo, even though Basir did not
do that (he added the file to the .gitignore file) because in the Node.js docs
they say you should put this file in the repo, so that others working with the
code all have the same lock file.