Lesson 9: Create Node/Express server on backend.
We transformed the entire root 'amazona' folder into a node package with the
command 'npm init'. This creates the package.json manifest file in the root.
Inside package.json file, add "type": "module" so node will treat javascript
code as ES 6 and understand ES modules and import/export statements.

Install Express via npm install express. Create a basic server.js file, with
routes for '/' and '/api/products'. We moved the data.js file from the frotend
to the backend folder so we can import it in the node server.

We installed the node monitor nodemon for hot reloading of the server when we
edit source files. npm install --save-dev nodemon. We also defined a 'start'
script in the package.json file so we can easily run the complicated command to
start the node server with nodemon watching the backend folder for changes.