Lesson 10: Change frontend react app to load products from back end.
Added 'proxy' setting to the package.json file in frontend folder pointing to
127.0.0.1:5000. In the frontend folder/package, used npm to install axios
library.

Edited the HomeScreen to define react useState hooks for variables products,
loading and error. Defined useEffect hook with a function to fetch the data from
the backend Node/Express server, using a GET ajax request to /api/products.

We created two new react components, LoadingBox and MessageBox. LoadingBox
displays a Loading... message with a spinner (pulled from font-awesome) while
the ajax request is sent to the backend server. MessageBox displays a nicely
formatted message either informational (in blue color) or indicating danger (in
red color). In the HomeScreen, if the ajax request for the product list results
in an error, we display it in a red MessageBox.

Also, I copied the data.js file that I had moved from frontend/src/ to backend/
back to frontend/src/ directory, because the ProductScreen component still
imports it for its data. The code will not compile without this file being
present on the frontend, until we change ProductScreen to get its data from the
backend as well, like HomeScreen.