Lesson 13: Add Redux to product details screen.
Created product details constants (PRODUCT_DETAILS_REQUEST, _SUCCESS, _FAIL),
actions (detailsProduct() function in productActions.js), and
reducers (productDetailsReducer() function in productReducers.js).

In the Redux store (store.js), add 'productDetails: productDetailsReducer' to
the object passed to combineReducers, to create the overall reducer of the
store.

Modified ProductScreen.js to use our new detailsProduct() action. We dispatch
detailsProduct(productId) in the useEffect hook, and get productDetails via
useSelector(state => state.productDetails), which I believe uses the
productDetailsReducer to yield an object which has loading, error and product
properties. We added <LoadingBox /> and <MessageBox /> components to display a
spinner while data is being fetched from the back end, and to format nicely any
error returned using the <MessageBox /> component.

In server.js, add a new handler for the '/api/product/:id' route on the backend,
which finds the individual product with the product id requested from the
data.js file (here on the backend server, now) and returns it to the frontend,
or sends a 404 error with message if that product id is not found.