Lesson 14: Handle Add to Cart Button
Created new CartScreen.js in frontend/src/screens/ directory, and in App.js we
added a new <Route /> with a path of '/cart/:id?' and set it to be handled by
the CartScreen component.

In the ProductScreen.js, we changed the last list item on the right hand column
which displayed the 'Add to Cart' button, to use conditional rendering to only
render when product.countInStock > 0. We also added a new line for selecting the
quantity to add to the cart.

Also made some style changes in index.css. We made the Qty select element have
similar styling to buttons, except that it does not change the mouse cursor to a
pointer (that should only be done for buttons and links).