Lesson 15: Implement Add to Cart action.
We created an action function and a reducer function for adding to the shopping
cart, along with a constant to identify the CART_ADD_ITEM action.

We modified the CartScreen to use our new action to add an item to the cart.

We added a badge to the shopping cart in the header, indicating number items in
the cart.

We used local storage to store the cart items, so that they are not lost if the
user refreshes the browser, or closes and re-opens it, the cart items will
remain.