Lesson 16: Build the nicer Shopping Cart screen
In CartScreen.js, we return JSX to create 2 columns: cart items on the left and
a cart action section on the right with the 'Proceed to Checkout' button.

We added a function to handle clicking of the 'Proceed to Checkout' button,
which redirects to the sign in screen, shipping section.

We add 'Delete' button to each item in the cart list, so user can remove that
item from the cart. We just added a stub function to handle this delete item
from cart action. We will implement this in the next lesson.