Lesson 17: Implement Remove from Cart action
Created a new constant for the remove item from cart action, CART_REMOVE_ITEM.

Implemented the function removeFromCart() for the action of removing an item
from the cart. This dispatches the CART_REMOVE_ITEM action with payload equal to
the product id of the item to remove from the cart.

Modified cartReducer function to handle the case of the CART_REMOVE_ITEM action,
which returns a copy of the state object with cartItems property updated to
filter out the item user wants to delete from the list.

In the CartScreen, modified the removeFromCartHandler() function, so that it
dispatches a call to removeFromCart().