Lesson 19: Create Sample Products in the Mongo DB
Created backend/models/productModel.js file, containing code to implement a
product schema through mongoose and a product model.

Created a productRouter.js file, with code to implement a productRouter that is
mounted at /api/products, so that it will have handlers for all routes that
match /api/products/... Added a get handler for '/seed', which inserts the list
of products from the data.js file.

Also added other productRouter handlers for routes '/' and '/:id'. The '/'
handler gets all documents from the Mongo database products collection, and
sends them to the frontend. The '/:id' handler finds the one product with id
matching the id in the URL, and if found sends it to the frontend, otherwise
returns 404 status code with error message 'Product not found.'

Opened the URL /api/products/seed to seed the Mongo database with the sample
products from the data.js file. We saw how if after doing this we refresh the
page, we get an error regarding the attempt to create documents with duplicate
name fields. We specified that the name field must be unique in our Schema.
Played around with adding code to delete all products from the database before
seeding, so we can refresh the page repeatedly, and it empties the products
collection and re-seeds with a new copy of the products from data.js file.
Finally, commented out the line that deletes all products, as it is a security
hole. We don't want anyone to get /api/products/seed and delete all our
products!