Lesson 20: Create Sign-in Backend API endpoint.
In userRouter.js, add a handler function for the endpoint '/api/users/signin'.
Since the routes in this file are mounted at '/api/users', this handler is
registered for the path '/signin'. Since this function will handle POST requests
with submitted email and password in the request body, we use userRouter.post().
In this function, bcrypt.compareSync() is used to check if the submitted
password matches the hashed password stored in the database. If the submitted
email and password are correct, then we generate a token using a generateToken()
function we created in the utils.js file in backend/.

The generateToken() function uses jwt.sign() function imported from the
jsonwebtoken package we installed. This function requires a secret to be passed
in, which is used to sign/encrypt the user's information. This secret must be
kept securely, so we store it in the file .env in the root folder. The
alternative value which JWT_SECRET is set to if it cannot be found in
process.env is there for the case where someone downloads the source code, which
does not include the .env file, so they won't have the secret. This provides a
secret for them so the code will run without error. However, in production, one
must use the secret contained in the .env file.