Lesson 21: Design Sign-In Screen
Created SigninScreen.js in frontend/src/screens/ to implement the sign-in
screen. Added a new route to App.js for '/signin' and set its handler to
our new SigninScreen component.

In our SigninScreen component, we added a stub function submitHandler()
which will eventually be used to handle clicking the 'Sign In' button,
which is the submit button of the form in the SigninScreen component.

Added some CSS to frontend/src/index.css style the sign-in screen.